+++ src/main/java/com/will/weather/dto/ForecastView.java
@@ -7,7 +7,7 @@ public record ForecastView(
         BigDecimal latitude,
         double temp,
         double feelsLike,
-        double humidity,
+        int humidity,
         String main,
         String description,
         String country,
